+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # RSYNC-GUI Changelog
 
+## 1.0.5
+- Switched to PyQt6 for macOS Linux cross-platform compatibility.
+- GUI, UX improvements. 
+- Added .deb package for Debian based Linux.
+
 ## 1.0.3
 - Now executes commands in terminal.
 - Major bookmarks rework! Now saves complete jobs and are editable.